requirement.txt
accelerate==1.1.1
adlfs==2023.12.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.6.2
appnope==0.1.4
asgiref==3.5.2
asttokens==2.2.0
async-timeout==4.0.3
attrs==22.1.0
Authlib==1.2.0
azure-common==1.1.27
azure-core==1.18.0
azure-datalake-store==0.0.53
azure-identity==1.6.1
azure-keyvault==4.1.0
azure-keyvault-certificates==4.3.0
azure-keyvault-keys==4.4.0
azure-keyvault-secrets==4.3.0
azure-nspkg==3.0.2
azure-storage-blob==12.8.1
azure-storage-queue==12.2.0
backcall==0.2.0
cached-property==1.5.2
cachetools==5.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.0.7
cryptography==3.4.8
cssselect==1.2.0
cssutils==2.6.0
ctransformers==0.2.27
cycler==0.11.0
dacite==1.8.1
dask==2021.10.0
dataclasses-json==0.6.7
datasets==3.1.0
debugpy==1.6.4
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
docker-pycreds==0.4.0
elastic-transport==8.4.0
elasticsearch==7.16.2
email-to==0.1.0
entrypoints==0.4
environs==9.5.0
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.88.0
filelock==3.16.1
fonttools==4.38.0
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.1.1
future==0.18.2
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.11.0
google-auth==2.15.0
googleapis-common-protos==1.57.0
GQLAlchemy==1.6.0
greenlet==3.0.3
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.2
importlib-metadata==5.1.0
importlib-resources==5.10.2
ipykernel==6.17.1
ipython==8.7.0
ipywidgets==8.1.5
isodate==0.6.0
jedi==0.18.2
Jinja2==3.1.3
joblib==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.4
kor==3.0.0
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.1.147
llama_cpp_python==0.3.2
locket==1.0.0
loguru==0.6.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
Metaphone==0.6
milvus-lite==2.4.10
mime==0.1.0
mpmath==1.3.0
msal==1.15.0
msal-extensions==0.3.0
msgpack==1.1.0
msrest==0.6.21
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
neo4j==4.4.12
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.25.2
oauthlib==3.1.1
opencensus==0.11.0
opencensus-context==0.1.3
opencensus-ext-azure==1.1.7
opencv-python==4.10.0.84
openpyxl==3.0.10
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.3
partd==1.4.2
pdf2image==1.17.0
peft==0.13.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.4
portalocker==1.7.1
premailer==3.10.0
prompt-toolkit==3.0.33
propcache==0.2.0
protobuf==4.21.10
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.13.0
PyJWT==2.2.0
pymgclient==1.3.1
pymilvus==2.4.9
pymongo==3.12.0
pymssql==2.3.0
PyMySQL==1.0.2
pyodbc==4.0.35
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.1
PyYAML==6.0.2
pyzmq==24.0.1
ray==2.37.0
redmail==0.6.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
rpds-py==0.20.0
rsa==4.8
safetensors==0.4.5
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.13.2
secure-smtplib==0.1.1
sentry-sdk==2.19.0
setproctitle==1.3.4
sharepy==2.0.0
shortuuid==1.0.11
simple-salesforce==1.12.2
six==1.16.0
skfeature-chappers==1.1.0
sklearn-features==0.0.2
smmap==5.0.1
sniffio==1.2.0
SQLAlchemy==1.4.44
stack-data==0.6.2
starlette==0.22.0
sweetviz==2.3.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.1.0
tokenizers==0.20.3
toolz==0.12.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.2
tqdm==4.66.4
traitlets==5.6.0
transformers==4.46.3
trino==0.328.0
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
ultralytics==8.3.34
ultralytics-thop==2.0.12
urllib3==2.2.3
uvicorn==0.20.0
wandb==0.18.7
wcwidth==0.2.5
widgetsnbextension==4.0.13
win32-setctime==1.1.0
xgboost==2.0.3
XlsxWriter==3.0.3
xlwt==1.3.0
xxhash==3.5.0
yarl==1.14.0
zeep==4.2.1
zipp==3.7.0
